Don't rely on the text from getqflist
The text for the items on the quickfix window isn't always the content
of the buffer. It can be other things, like error messages or the path
to the file. To make replacer.nvim work with any quickfix window we need
to ignore the text and simply read the line from the file.

Fixes https://github.com/gabrielpoca/replacer.nvim/issues/10